Use Glob package for globbing instead
As filemanip suffers from https://github.com/bos/filemanip/pull/10

Glob conveniently provides ‘globDir’ which expands all the patterns for
us but it does not provide the list of patterns which did not match
anything. To do that, we ‘globDir’ with just one pattern and then
‘match’ on all files it has encountered: this approach means we still
only traverse the tree once.