Exit successfully in more cases.
Previously, `hpack` would return with a failure exit code if it
detected that the `.cabal` file had been previously generated by a
newer version of `hpack` - even if the `.cabal` file was precisely
what would have been generated anyway!

This is an usability improvement usability whenever there are multiple
versions of `hpack` in use in a single environment (e.g., one version
baked into `stack`, and a stand-alone version being used for
experimentation and evaluation of new features).