Allow dependency modules to use mixins.
When used, this will produce `mixins` fields in the generated package
description. It doesn't do anything too clever: it just gloms the list
of strings onto the package name.

Along the way, defining individual dependencies as objects has become
more flexible: they can now be a objects with `name` and `version`
keys. Some of this was present before due to `SourceDependency`, but
now it is more flexible and is useful generally, rather than only for
integration with tinc.

See #318.