Don't write out the cabal file if it matches the old one
The motivation behind this change is avoid the `--force` option from triggering
an unnecessary reload of `ghcid` if nothing changed.

The context is that at work we use `hpack --force` to correct manual edits
that users might make to the `.cabal` file, and we run `hpack --force`
whenever users enter our Nix shell.  If a user uses `direnv` on top of that
then `hpack --force` runs every time they `cd` to the project directory.

However, suppose that they load `ghcid` in one window and then `cd` to the
project directory in another window.  This causes `hpack --force` to run and
even if nothing changed the `mwb.cabal` file is touched which causes `ghcid`
to reload (which is expensive).  This change fixes that.